requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
openai==1.77.0
requests==2.32.3
anthropic==0.51.0